feat(cache-module): added utils
