+++ package.json
@@ -46,6 +46,7 @@
     ]
   },
   "dependencies": {
+    "@americanexpress/lumberjack": "^1.1.4",
     "compression": "^1.7.4",
     "cors": "^2.8.4",
     "got": "^11.8.3",
